Move filter points by distance from line to annotation_operators
Did this so that filtering points by their distance from a line
annotation can happen outside of the FilterPointDistanceFromRow
ToolboxItem.
Also its a function that operates on annotations, so I think that it
makes sense for it to be there.
